% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_retro_score}
\alias{calculate_retro_score}
\title{Calculate retro-scores for all groups}
\usage{
calculate_retro_score(
  df,
  col_date_occurrence,
  col_date_reporting,
  col_value,
  group_cols = NULL,
  method = "2D_allgroups",
  max_delay = Inf,
  aggrby
)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{col_date_occurrence}{Column name for the date of occurrence/reference.}

\item{col_date_reporting}{Column name for the date of reporting.}

\item{col_value}{Column name for the value.}

\item{group_cols}{Optional character vector of column names for grouping.}

\item{method}{'2D_allgroups' (number of changes in 2D triangle) or 'at_least_1_change_by_occ' (number of occurrence dates with at least 2 reported values)}

\item{max_delay}{Maximum delay to consider. (only works with method '2D_allgroups')}

\item{aggrby}{A character vector of column names to aggregate by.}
}
\value{
A tibble with group cols + retro_score (percentage 0-1)
}
\description{
The retro-score is the amount of retro-adjustments / max possible retro-adjustments
The higher the better for nowcast_cl()
retro_score = n_changes / max_changes or = retro_adjustments / max_retro_adj
Notes:
"retro-adjustments" = "value changes"
retro-score = number of changes / number of ywks (max changes)
}
\examples{
generate_test_data() \%>\%
  calculate_retro_score(
    col_date_occurrence = date_occurrence,
    col_date_reporting = date_report,
    col_value = value,
    group_cols = NULL
    # , aggrby = country
    # , method = "at_least_1_change_by_occ"
  )

}
