% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_nowcast.R
\name{explore_nowcast}
\alias{explore_nowcast}
\title{Explore Nowcast Results with Shiny}
\usage{
explore_nowcast(nc_obj)
}
\arguments{
\item{nc_obj}{A \code{nowcast_results} S7 object.}
}
\description{
Explore Nowcast Results with Shiny
}
