% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_future_reported_values}
\alias{fill_future_reported_values}
\title{Fill future reported values with last known values}
\usage{
fill_future_reported_values(
  df,
  col_date_occurrence,
  col_date_reporting,
  col_value,
  group_cols = NULL,
  max_delay = Inf
)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{col_date_occurrence}{Column name for the date of occurrence/reference.}

\item{col_date_reporting}{Column name for the date of reporting.}

\item{col_value}{Column name for the value.}

\item{group_cols}{Optional character vector of column names for grouping.}

\item{max_delay}{Inf / 'auto' / NULL / integer. 'auto' or NULL will keep the same max_delay as the input.}
}
\value{
A data frame with the same columns as \code{df}, but with rows added for
missing reporting dates and \code{NA} values in \code{col_value} filled with the
last available observation for each occurrence date within each group.
}
\description{
This function completes a data frame to include all combinations
of occurrence and reporting dates (within each group).
It then fills in missing values by carrying the last known reported value
backward in time from future reports to past reports for a given occurrence.
This is useful for dealing with right-censored reporting data where reports
are updated over time.
}
\examples{
library(dplyr)

generate_test_data() \%>\%
  fill_future_reported_values(
    col_date_occurrence = date_occurrence,
    col_date_reporting = date_report,
    col_value = value,
    group_cols = NULL
  )
}
