% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nowcast_demo}
\alias{nowcast_demo}
\title{Nowcast Dataset for demo and testing}
\format{
A data frame with X rows and Y variables:
\describe{
\item{group}{text, for grouping}
\item{date_occurrence}{text, representing week of onset}
\item{date_report}{text, representing week of reporting}
\item{value}{number, representing number of cases}
}
}
\usage{
nowcast_demo
}
\description{
This dataset contains data for demonstration purposes.
}
\keyword{datasets}
