% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_cl.R
\name{nowcast_results}
\alias{nowcast_results}
\title{S7 object class for \code{nowcast_cl()} Results}
\usage{
nowcast_results(name, params, time_start, time_end, n_groups, max_delay,
  data, completeness, delays, models, results)
}
\arguments{
\item{name}{A character string with a timestamp for the run.}

\item{params}{A list with the parameters used for the nowcasting (unevaluated call).}

\item{time_start}{the sys time at which the function started.}

\item{time_end}{the sys time at which the function ended.}

\item{n_groups}{The number of groups processed.}

\item{max_delay}{The maximum delay used.}

\item{data}{Dataframe. The original input data frame (with only required columns).}

\item{completeness}{Dataframe. The original input data frame with delays and completeness columns.}

\item{delays}{Dataframe. A data frame with the final aggregated completeness estimates per delay (+ \code{modelled} column if do_model_fitting was TRUE).}

\item{models}{Dataframe. The resulting fitted models (empty data frame if do_model_fitting was FALSE)}

\item{results}{Dataframe. A data frame with the resulting nowcasting predictions.}
}
\value{
An \bold{S7 object} of class \code{nowcast_results}. This object is a
structured container for the entire nowcasting pipeline output. It consists
of the following properties (slots):
\describe{
\item{name}{Character. A unique timestamp identifier for the run (\code{YYYYMMDD_HHMMSS}).}
\item{params}{List. The evaluated parameters and arguments used in the function call.}
\item{time_start, time_end}{POSIXct. Timestamps marking the duration of the calculation.}
\item{n_groups}{Numeric. The total count of unique groups processed.}
\item{max_delay}{Numeric. The maximum reporting delay (in \code{time_units}) considered.}
\item{data}{Data frame. The subset of the original input used for the analysis.}
\item{completeness}{Data frame. Detailed row-level completeness calculations and delays.}
\item{delays}{Data frame. Aggregated completeness estimates per delay unit, including
both observed and (optionally) modelled values.}
\item{models}{Data frame. Results of the non-linear model fitting, including RSS and
model types. Returns an empty data frame if \code{do_model_fitting} was \code{FALSE}.}
\item{results}{Data frame. The final nowcasting table containing predicted values.}
}
}
\description{
The \code{nowcast_cl()} function returns an object of this class.
}
