% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nowcast_test_data}
\alias{nowcast_test_data}
\title{Nowcast Minimal Dataset for testing}
\format{
A data frame with X rows and Y variables:
\describe{
\item{country}{text, for grouping}
\item{onset_week}{text, representing week of onset}
\item{report_week}{text, representing week of reporting}
\item{cases}{number, representing number of cases}
}
}
\usage{
nowcast_test_data
}
\description{
This dataset contains data for testing purposes.
}
\keyword{datasets}
