% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.nowcast_results}
\alias{plot.nowcast_results}
\title{Plotting method for nowcasting results}
\usage{
\method{plot}{nowcast_results}(x, which, option, do_rescale, add_model_info, ...)
}
\arguments{
\item{x}{A \code{nowcast_results} object.}

\item{which}{Which plot to draw, one of "results", "data", or "delays".}

\item{option}{Either "millipede" or "triangle". Only for \code{which = "data"}.}

\item{do_rescale}{Rescale values 0-1. Boolean. Only for \code{which = "data"} or \code{which = "delays"}.}

\item{add_model_info}{Add model info to the plot. Boolean. Only for \code{which = "delays"}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A ggplot object.
}
\description{
This is a plot method which can plot all \code{nowcast_cl()} outputs:
\itemize{
\item data (\code{which = "data"})
\item delays (\code{which = "delays"})
\item results (\code{which = "results"})
It will apply a theme, a caption and a facet wrap on groups.
}
}
