% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_delays}
\alias{plot_delays}
\title{Plot Reporting Completeness by Delay}
\usage{
plot_delays(
  df,
  col_completeness_obs,
  col_completeness_modelled = "",
  group_cols = NULL,
  color1 = "#222222",
  color2 = "firebrick2",
  limits_y = c(NA, NA)
)
}
\arguments{
\item{df}{A data.frame containing 'delay' and \code{col_completeness} columns.
An optional \code{modelled} column}

\item{col_completeness_obs}{Column name for the Observed Completeness. (dots)}

\item{col_completeness_modelled}{Column name for the Modelled Completeness. (line)}

\item{group_cols}{Optional character vector of column names for grouping.}

\item{color1}{Color for observed data. (dots)}

\item{color2}{Color for modelled data. (line)}

\item{limits_y}{vector to be passed to limits of \code{ggplot2::scale_y_continuous}.}
}
\value{
A ggplot object showing completeness vs. delay.
}
\description{
Creates a scatter plot of reporting completeness against reporting delay.
If a \code{col_completeness_modelled} is present, it will be shown as a dotted line.
}
\examples{
delays <- data.frame(
  delay = 0:9,
  completeness = c(0.509, 0.802, 0.920, 0.967, 0.987, 0.995, 0.998, 0.999, 1, NA),
  modelled = c(0.509, 0.802, 0.920, 0.968, 0.987, 0.995, 0.998, 0.999, 1, 1)
)
plot_delays(
  df = delays,
  col_completeness_obs = completeness,
  col_completeness_modelled = modelled
)

}
