% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_millipede}
\alias{plot_millipede}
\title{Millipede plot}
\usage{
plot_millipede(
  df,
  col_value,
  col_date_occurrence,
  col_date_reporting,
  scale_percent = FALSE
)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{col_value}{Column name for the value.}

\item{col_date_occurrence}{Column name for the date of occurrence/reference.}

\item{col_date_reporting}{Column name for the date of reporting.}

\item{scale_percent}{Logical; if TRUE formats the fill legend as percentages and labels it "Percentage".}
}
\value{
A ggplot object.
}
\description{
Draw a line plot where each line represents one reporting date:
\itemize{
\item x = date of occurrence
\item y = value (e.g., counts or proportions)
\item group + fill = date of reporting
}
}
\details{
When there are reporting delays the plot look like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{---\\--\\--\\--\\
    \\  \\  \\  \\
     \\  \\  \\  \\
}\if{html}{\out{</div>}}
}
\examples{
generate_test_data() \%>\%
  plot_millipede(
    col_value = value,
    col_date_occurrence = date_occurrence,
    col_date_reporting = date_report,
    scale_percent = FALSE
  )

}
\seealso{
ggplot2::geom_line, ggplot2::scale_color_viridis_c
}
