% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nc_input}
\alias{plot_nc_input}
\title{Plot Nowcast Input Data}
\usage{
plot_nc_input(
  df,
  col_value,
  col_date_occurrence,
  col_date_reporting,
  group_cols = NULL,
  option = "millipede",
  do_rescale = TRUE,
  do_facet_groups = TRUE
)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{col_value}{Column name for the value.}

\item{col_date_occurrence}{Column name for the date of occurrence/reference.}

\item{col_date_reporting}{Column name for the date of reporting.}

\item{group_cols}{Optional character vector of column names for grouping.}

\item{option}{"millipede" or "triangle".}

\item{do_rescale}{Rescale values 0-1. Boolean.}

\item{do_facet_groups}{Boolean. Should groups be faceted?}
}
\value{
A ggplot object.
}
\description{
Can plot 2 types of plot: option="triangle" or "millipede"
}
