% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_repeated_values}
\alias{rm_repeated_values}
\title{Remove duplicated reported values in reporting matrix}
\usage{
rm_repeated_values(
  df,
  col_date_occurrence,
  col_date_reporting,
  col_value,
  group_cols = NULL
)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{col_date_occurrence}{Column name for the date of occurrence/reference.}

\item{col_date_reporting}{Column name for the date of reporting.}

\item{col_value}{Column name for the value.}

\item{group_cols}{Optional character vector of column names for grouping.}
}
\value{
A tibble with the same columns as \code{df}, but with rows removed.
}
\description{
Remove duplicated reported values in reporting matrix
}
\examples{
library(dplyr)
generate_test_data(n_delays = 20, n_reportdates = 20) \%>\%
  mutate(value = round(value, 1)) \%>\% ## make values identical
  rm_repeated_values(
    col_value = value,
    col_date_occurrence = date_occurrence,
    col_date_reporting = date_report
  )
}
