% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_eval.R
\name{nowcast_eval_results}
\alias{nowcast_eval_results}
\title{S7 object class for \code{nowcast_eval()} Results}
\usage{
nowcast_eval_results(
  detail,
  summary,
  params,
  n_past,
  time_start,
  time_end
)
}
\arguments{
\item{detail}{data.frame. Per-prediction errors with columns for observed value,
predicted value, true value.}

\item{summary}{data.frame. Aggregated metrics per group x delay:
SMAPE (pred and obs), SMAPE improvement, proportion_pred_is_better, Wilson CIs.}

\item{params}{list. Parameters used for the evaluation run.}

\item{n_past}{integer. Number of past reporting periods evaluated.}

\item{time_start}{POSIXct. Time the function started.}

\item{time_end}{POSIXct. Time the function ended.}
}
\value{
An S7 object of class \code{nowcast_eval_results} with the following slots:
\describe{
\item{detail}{Data frame. Per-prediction errors with columns for observed value,
predicted value, true value.}
\item{summary}{Data frame. Aggregated metrics per group x delay:
SMAPE (pred and obs), SMAPE improvement, proportion_pred_is_better, Wilson CIs.}
\item{params}{List. Parameters used for the evaluation run.}
\item{n_past}{Numeric. Number of past reporting periods evaluated.}
\item{time_start}{POSIXct. Time the function started.}
\item{time_end}{POSIXct. Time the function ended.}
}
}
\description{
The \code{nowcast_eval()} function returns an object of this class.
}
