% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tbl_models_stats}
\alias{tbl_models_stats}
\title{Unwrap nowcast models stats}
\usage{
tbl_models_stats(nc_obj, thresholds_r2 = 0.8)
}
\arguments{
\item{nc_obj}{A \code{nowcast_results} object.}

\item{thresholds_r2}{R squared threshold to classify \code{eval} into good or bad fit}
}
\value{
A tibble.
}
\description{
Unwrap nowcast models stats
}
